Move the DummyClass GodotResource to the correct package
