+++ godot-kotlin/godot-library/src/nativeGen/kotlin/godot/GodotResource.kt
@@ -1,4 +1,4 @@
-package godot.core
+package godot
 
 //TODO: this needs to be properly implemented. It's just here for the implementation of the PropertyTypeHintAnnotations
 class GodotResource {
